+++ CHANGELOG.rst
@@ -727,6 +727,9 @@ Bug Fixes
 * Fix URL of TYNDP scenario dataset
 * Automatically generated tasks now get unique :code:`task_id`\s.
   Fixes issue `#985`_ via PR `#986`_.
+* Set upper limit on commissioning date for units from MaStR
+  dataset
+  `#1098 <https://github.com/openego/eGon-data/issues/1098>`_
 
 .. _PR #692: https://github.com/openego/eGon-data/pull/692
 .. _#343: https://github.com/openego/eGon-data/issues/343